1st (big?, 6th of December 2024) Update To PythonScript
Added more features, but still not quite satisfied. I think I might add a "merge at junction" feature like I did with another addon of mine which I havn't released cause its literally just this addon, but for walls.

P.S. junction is the word for a meeting point between walls & streets & other things. A 4 way junction in this case, is just a wall object, but with 4 paths extending from a slim'ish wall, the slim wall being slim cause its not extending a whole meter with just one side of itself being extended... Of course, the reason for the junction being split the way it is, is because I wanted to maintain spacing between the wall-objects, I'm hoping to add verts into this block - placing addon at some point too btw, but I'll be doing it manually & I don't want to confuse claude AI by including it in this script so... yeah ;p